+++ src/pitch.cpp
@@ -13,17 +13,12 @@ Pitch::Pitch(const std::string& tag, const std::string& schema)
     auto output_gain = gst_element_factory_make("volume", nullptr);
     auto out_level = gst_element_factory_make("level", "pitch_output_level");
     auto audioconvert_in = gst_element_factory_make("audioconvert", nullptr);
-    auto audioresample_in = gst_element_factory_make("audioresample", nullptr);
-    auto audiorate = gst_element_factory_make("audiorate", nullptr);
     auto audioconvert_out = gst_element_factory_make("audioconvert", nullptr);
-    auto audioresample_out = gst_element_factory_make("audioresample", nullptr);
 
-    gst_bin_add_many(GST_BIN(bin), input_gain, in_level, audioconvert_in, audioresample_in,
-                     pitch, audiorate, audioconvert_out, audioresample_out, output_gain,
-                     out_level, nullptr);
-    gst_element_link_many(input_gain, in_level, audioconvert_in, audioresample_in, pitch,
-                          audiorate, audioconvert_out, audioresample_out, output_gain,
-                          out_level, nullptr);
+    gst_bin_add_many(GST_BIN(bin), input_gain, in_level, audioconvert_in, pitch,
+                     audioconvert_out, output_gain, out_level, nullptr);
+    gst_element_link_many(input_gain, in_level, audioconvert_in, pitch,
+                          audioconvert_out, output_gain, out_level, nullptr);
 
     auto pad_sink = gst_element_get_static_pad(input_gain, "sink");
     auto pad_src = gst_element_get_static_pad(out_level, "src");